Some file lock attempting...

+++ WoWFormatLib/FileProviders/WagoFileProvider.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net.Http;
+using System.Threading;
 
 namespace WoWFormatLib.FileProviders
 {
@@ -11,6 +12,8 @@ namespace WoWFormatLib.FileProviders
         private HttpClient Client = new();
         private string Build;
         private Dictionary<uint, string> Files = [];
+        private Lock BuildLock = new();
+        private Dictionary<uint, Lock> FileLocks = [];
 
         public void SetBuild(string build)
         {
@@ -23,23 +26,33 @@ namespace WoWFormatLib.FileProviders
                 return;
             }
 
-            Console.WriteLine("Fetching file list from Wago");
-            var response = Client.GetAsync("https://wago.tools/api/files?version=" + Build).Result;
-
-            if (response.IsSuccessStatusCode)
+            lock (BuildLock)
             {
-                var fileList = response.Content.ReadAsStringAsync().Result;
+                if (File.Exists("cache/wago/" + Build + ".json"))
+                {
+                    Console.WriteLine("Reading wago file list from cache");
+                    Files = System.Text.Json.JsonSerializer.Deserialize<Dictionary<uint, string>>(File.ReadAllText("cache/wago/" + Build + ".json"));
+                    return;
+                }
 
-                if (!Directory.Exists("cache/wago/files"))
-                    Directory.CreateDirectory("cache/wago/files");
+                Console.WriteLine("Fetching file list from Wago");
+                var response = Client.GetAsync("https://wago.tools/api/files?version=" + Build).Result;
 
-                File.WriteAllText("cache/wago/" + Build + ".json", fileList);
+                if (response.IsSuccessStatusCode)
+                {
+                    var fileList = response.Content.ReadAsStringAsync().Result;
 
-                Files = System.Text.Json.JsonSerializer.Deserialize<Dictionary<uint, string>>(fileList);
-            }
-            else
-            {
-                throw new Exception("Failed to fetch file list from Wago");
+                    if (!Directory.Exists("cache/wago/files"))
+                        Directory.CreateDirectory("cache/wago/files");
+
+                    File.WriteAllText("cache/wago/" + Build + ".json", fileList);
+
+                    Files = System.Text.Json.JsonSerializer.Deserialize<Dictionary<uint, string>>(fileList);
+                }
+                else
+                {
+                    throw new Exception("Failed to fetch file list from Wago");
+                }
             }
         }
 
@@ -55,33 +68,34 @@ namespace WoWFormatLib.FileProviders
 
         public Stream OpenFile(uint filedataid)
         {
-            if (Files.ContainsKey(filedataid))
-            {
-                if (!Directory.Exists("cache/wago/files/" + Build))
-                    Directory.CreateDirectory("cache/wago/files/" + Build);
+            if (!Directory.Exists("cache/wago/files/" + Build))
+                Directory.CreateDirectory("cache/wago/files/" + Build);
 
-                if (File.Exists("cache/wago/" + Build + "/" + filedataid))
-                    return File.OpenRead("cache/wago/files/" + Build + "/" + filedataid);
+            if (File.Exists("cache/wago/files/" + Build + "/" + filedataid))
+                return new MemoryStream(File.ReadAllBytes("cache/wago/files/" + Build + "/" + filedataid));
+
+            if (!FileLocks.ContainsKey(filedataid))
+                FileLocks[filedataid] = new Lock();
+
+            lock(FileLocks[filedataid])
+            {
+                // Check again if the file exists after acquiring the lock
+                if (File.Exists("cache/wago/files/" + Build + "/" + filedataid))
+                    return new MemoryStream(File.ReadAllBytes("cache/wago/files/" + Build + "/" + filedataid));
 
                 var response = Client.GetAsync("https://wago.tools/api/casc/" + filedataid + "?version=" + Build).Result;
 
                 if (response.IsSuccessStatusCode)
                 {
                     using (var fileStream = File.Create("cache/wago/files/" + Build + "/" + filedataid))
-                    {
                         response.Content.CopyToAsync(fileStream).Wait();
-                    }
-                        
-                    return File.OpenRead("cache/wago/files/" + Build + "/" + filedataid);
                 }
                 else
                 {
                     throw new Exception("Failed to fetch file " + filedataid + " from Wago");
                 }
-            }
-            else
-            {
-                throw new FileNotFoundException("File with filedataid " + filedataid + " not found");
+
+                return new MemoryStream(File.ReadAllBytes("cache/wago/files/" + Build + "/" + filedataid));
             }
         }
 
@@ -89,7 +103,6 @@ namespace WoWFormatLib.FileProviders
         {
             throw new NotImplementedException();
         }
-
         public uint GetFileDataIdByName(string filename)
         {
             return Files.FirstOrDefault(x => x.Value == filename).Key;
